+++ adapters/IronSource/IronSourceAdapter/GADMAdapterIronSourceConstants.h
@@ -17,8 +17,10 @@ static NSString *const GADMAdapterIronSourceMediationName = @"AdMob";
 
 static NSString *const GADMAdapterIronSourceAdapterVersion = @"8.3.0.0.0";
 
-static NSString *const GADMIronSourceDefaultInstanceId = @"0";
-static NSString *const GADMAdapterIronSourceInternalVersion = @"410";
+static NSString *const GADMIronSourceDefaultNonRtbInstanceId = @"0";
+static NSString *const GADMIronSourceDefaultRtbInstanceId = @"";
+
+static NSString *const GADMAdapterIronSourceInternalVersion = @"500";
 
 // IronSource parameters keys.
 static NSString *const GADMAdapterIronSourceAppKey = @"appKey";
@@ -34,3 +36,5 @@ static InstanceState const GADMAdapterIronSourceInstanceStateShowing = @"SHOWING
 
 // IronSource mediation adapter error domain.
 static NSString *const GADMAdapterIronSourceErrorDomain = @"com.google.mediation.IronSource";
+
+static NSString *const GADMAdapterIronSourceWatermark = @"watermark";