+++ adapters/IronSource/IronSourceAdapter/GADMAdapterIronSourceUtils.h
@@ -46,7 +46,9 @@ NSError *_Nonnull GADMAdapterIronSourceErrorWithCodeAndDescription(
 
 + (nullable ISBannerSize *)ironSourceAdSizeFromRequestedSize:(GADAdSize)size;
 
-/// Sets the watermark using the ad configuration.
-+ (void)setWatermarkWithAdConfiguration:(nonnull GADMediationAdConfiguration *)adConfiguration;
++ (NSArray<ISAAdFormat *> *_Nullable)adFormatsToInitializeForAdUnits:(nonnull NSSet *)adUnits;
+
++ (nonnull NSMutableDictionary<NSString *, NSString *> *)getExtraParamsWithWatermark:
+    (nullable NSData *)watermarkData;
 
 @end